test: fold agent/ micro-suites into their canonical module suites (#9253)
First tranche of #9253 scoped to src/test-kernel/agent/. Applies R7
(extend-first, one suite per product module): merges sub-150-line
micro-suites into the existing canonical suite for the same product
module, deleting the micro-suite file. Zero production changes.

- OutputFilesPrompt.vitest.mts + UserVarsMissingFiles.vitest.mts folded
  into src/test-kernel/agent/utils/userVars.vitest.ts (all three test
  src/agent/utils/userVars.ts). Merge order is deliberate: the folded
  "buildUserVars with missing configured files" describe replaces the
  global fake platform in its own beforeEach and never restores it, so
  it must run last, after the config-provider-based describes.
- AgentCreatorToolCatalogParity.vitest.mts folded into
  AgentCreatorOrchestration.vitest.ts (both test agentCreatorFlow.ts).
  The parity describe only reads the static TOOL_GROUPS export, so the
  host file's helperModel/agentLoad vi.mock()s do not affect it.

Net: 5 files -> 2 files, 26 test cases -> 26 test cases (same names,
verified via targeted vitest runs before and after).

Left unmerged after audit (documented in PR body): three suites that
share src/agent/storage/executionLifecycle.ts but use incompatible
module-level mocking (real store vs. fully-mocked ExecutionKVStore vs.
mocked agentRegistry); two suites sharing executionListing.ts with the
same real-vs-mocked-store conflict; and UserVarsDelegationScope.vitest.ts,
excluded from the userVars fold because its vi.mock('@agent/index/agentRegistry')
would leak into the other describes once merged into one file.

Refs #9253
